+++ src/components/transactions-table/components/transaction-hash/transaction-hash.tsx
@@ -11,9 +11,13 @@ export class TransactionHash {
   @Prop() transaction: ITransactionsTableRow;
 
   render() {
+    if (!this.transaction) {
+      return null;
+    }
+
     return (
       <div class="transactions-table-body-cell">
-        <transaction-icon icon-info={this.transaction.iconInfo}></transaction-icon>
+        <transaction-icon iconInfo={this.transaction.iconInfo}></transaction-icon>
         <explorer-link dataTestId={DataTestIdsEnum.transactionLink} link={this.transaction.link} text={this.transaction.txHash}></explorer-link>
       </div>
     );